Refactored SampleSlide into AppIntro. Changed AppIntroStyle to AppIntroTheme
