+++ appintro/src/main/res/values/styles.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <style name="AppIntroStyle" parent="Theme.AppCompat.DayNight.NoActionBar" />
+    <style name="AppIntroTheme" parent="Theme.AppCompat.DayNight.NoActionBar" />
 
     <style name="AppIntroButtonStyle" parent="Widget.AppCompat.Button.Borderless">
         <item name="android:background">?attr/selectableItemBackground</item>
