+++ example/src/main/AndroidManifest.xml
@@ -26,28 +26,28 @@
         <activity
             android:name="com.amqtech.opensource.appintroexample.ui.mainTabs.intro.DefaultIntro"
             android:label="@string/app_name"
-            android:theme="@style/AppIntroStyle" />
+            android:theme="@style/AppIntroTheme" />
         <activity
             android:name="com.amqtech.opensource.appintroexample.ui.mainTabs.intro.DefaultIntro2"
             android:label="@string/app_name"
-            android:theme="@style/AppIntroStyle" />
+            android:theme="@style/AppIntroTheme" />
         <activity
             android:name="com.amqtech.opensource.appintroexample.ui.mainTabs.intro.CustomLayoutIntro"
             android:label="@string/app_name"
-            android:theme="@style/AppIntroStyle" />
+            android:theme="@style/AppIntroTheme" />
         <activity
             android:name="com.amqtech.opensource.appintroexample.ui.mainTabs.intro.CustomBackgroundIntro"
             android:label="@string/app_name"
-            android:theme="@style/AppIntroStyle" />
+            android:theme="@style/AppIntroTheme" />
 
         <activity
             android:name="com.amqtech.opensource.appintroexample.ui.permsTabs.intro.PermissionsIntro1"
             android:label="@string/app_name"
-            android:theme="@style/AppIntroStyle" />
+            android:theme="@style/AppIntroTheme" />
         <activity
             android:name="com.amqtech.opensource.appintroexample.ui.permsTabs.intro.PermissionsIntro2"
             android:label="@string/app_name"
-            android:theme="@style/AppIntroStyle" />
+            android:theme="@style/AppIntroTheme" />
     </application>
 
 </manifest>
