+++ example/src/main/java/com/amqtech/opensource/appintroexample/ui/mainTabs/intro/CustomLayoutIntro.java
@@ -1,12 +1,13 @@
 package com.amqtech.opensource.appintroexample.ui.mainTabs.intro;
 
 import android.os.Bundle;
+
+import com.github.paolorotolo.appintro.AppIntro;
+import com.github.paolorotolo.appintro.CustomSlide;
+import com.github.paolorotolo.appintroexample.R;
+
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-
-import com.amqtech.opensource.appintroexample.util.SampleSlide;
-import com.github.paolorotolo.appintro.AppIntro;
-import com.github.paolorotolo.appintroexample.R;
 
 /**
  * Created by andrew on 11/17/16.
@@ -18,10 +19,10 @@ public class CustomLayoutIntro extends AppIntro {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        addSlide(SampleSlide.newInstance(R.layout.intro_custom_layout1));
-        addSlide(SampleSlide.newInstance(R.layout.intro_custom_layout2));
-        addSlide(SampleSlide.newInstance(R.layout.intro_custom_layout3));
-        addSlide(SampleSlide.newInstance(R.layout.intro_custom_layout4));
+        addSlide(CustomSlide.newInstance(R.layout.intro_custom_layout1));
+        addSlide(CustomSlide.newInstance(R.layout.intro_custom_layout2));
+        addSlide(CustomSlide.newInstance(R.layout.intro_custom_layout3));
+        addSlide(CustomSlide.newInstance(R.layout.intro_custom_layout4));
 
         setProgressIndicator();
     }
