+++ example/src/main/res/values/styles.xml
@@ -6,10 +6,4 @@
         <item name="colorAccent">#E91E63</item>
     </style>
 
-    <style name="AppIntroTheme" parent="MaterialTheme.TranslucentStatus">
-        <item name="colorPrimary">#1976D2</item>
-        <item name="colorPrimaryDark">#1565C0</item>
-        <item name="colorAccent">#E91E63</item>
-    </style>
-
 </resources>
